Refactor `BaseCommandOptions` import path
Moved `BaseCommandOptions` from general types to command types. This change affects the import paths in `langchain/utils.ts` and `commit/options.ts`. The import path has been updated to reflect the new location. This makes the code structure more logical and easier to understand.
